exception handler created and controllers bugs fixed

+++ src/main/java/com/levelvini/biblioteca/entities/Categoria.java
@@ -16,7 +16,7 @@ public class Categoria {
 
     @jakarta.persistence.Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer Id;
+    private Long Id;
     private String name;
 
     @Setter(AccessLevel.NONE)
